podman: WIP support podman native
To look into supporting direct container install, we allow the build
of podman as a native tool.

Note: This has issues running both under pseudo, some distros (missing
kernel features) and as non-root, but we still need this as a basic
building block feature.

local.conf needs:

  DISTRO_FEATURES_NATIVE:append = " seccomp"

For the build to succeed.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>